Session is now flushed on 'thanks'
The session for the current guest now gets flushed from the queue when the thanks
page laods. Also, a field for 'responded' was added to the party model allowing
tracking of whether someone from that party has completed the rsvp. Data is saved
on the confirm page, but the responded field is not filled in until 'thanks' loads.